feat: Update `postProcess` config to allow alternate save mechanism (#91)
Update `postProcess` configuration to run after tag clean-up phase
to allow an alternate mechanism for saving the generated Swagger to
be specified, rather than the standard writing a file to the filesystem.

Allow `swaggerJsonPath` to be empty/null/undefined, in which case
no file write attempted.

This supports deployment to environments that do not allow writing to the
filesystem e.g. store in memory or database.